# backend/xbank/swagger.yaml
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/hoaronal/XBANK/1.0.0
  description: SwaggerHub API Auto Mocking
tags:
- name: admins
  description: Secured Admin-only calls
- name: developers
  description: Operations available to regular developers
paths:
  /api/users/info:
    get:
      tags:
      - developers
      summary: users info
      operationId: usersInfo
      description: |
        Get user info
      responses:
        "200":
          description: user info
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: '#/components/schemas/UserDTO'
        "400":
          description: bad input parameter
    post:
      tags:
      - admins
      summary: adds an inventory item
      description: Adds an item to the system
      operationId: addInventory
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InventoryItem'
      responses:
        "201":
          description: item created
        "400":
          description: invalid input, object invalid
        "409":
          description: an existing item already exists
components:
  schemas:
    InventoryItem:
      required:
      - id
      - manufacturer
      - name
      - releaseDate
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: d290f1ee-6c54-4b01-90e6-d701748f0851
        name:
          type: string
          example: Widget Adapter
        releaseDate:
          type: string
          format: date-time
          example: 2016-08-29T09:12:33.001Z
        manufacturer:
          $ref: '#/components/schemas/Manufacturer'
    Manufacturer:
      required:
      - name
      type: object
      properties:
        name:
          type: string
          example: ACME Corporation
        homePage:
          type: string
          format: url
          example: https://www.acme-corp.com
        phone:
          type: string
          example: 408-867-5309
    UserDTO:
      type: object
      properties:
        id:
          type: string
          example: ACME Corporation
        login:
          type: string
          format: url
          example: https://www.acme-corp.com
        firstName:
          type: string
          example: admin
        lastName:
          type: string
          example: admin
        email:
          type: string
          example: admin@admin.com
        imageUrl:
          type: string
          example: https://cdn4.iconfinder.com/data/icons/small-n-flat/24/user-alt-512.png
        activated:
          type: boolean
          example: true
        langKey:
          type: string
          example: en
        createdBy:
          type: string
          example: admin
        createdDate:
          type: string
          example: 2020-12-12
        lastModifiedBy:
          type: string
          example: admin
        lastModifiedDate:
          type: string
          example: 2020-12-12